requirements.txt
Babel==2.7.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.13
django-coreapi==2.3.0
django-cors-headers==3.2.1
django-filter==2.3.0
django-phonenumber-field==3.0.1
django-tinymce==2.8.0
django-webpack-bundle==1.0.1
django-webpack-loader==0.6.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
phonenumbers==8.10.17
Pillow==6.1.0
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
